+++ src/app/project/view/project.view.component.css
@@ -24,36 +24,3 @@ h3 {
   font-size: 1.4em;
   padding-top: 0;
 }
-
-.share-card {
-  width: 70%;
-}
-
-.share-card .form {
-  margin: 0;
-}
-
-.share-card .email-preview {
-  display: block;
-}
-
-.share-card h2 {
-  margin: 0.5em 0 0.25em 0;
-  /*display: none;*/
-}
-
-.wide {
-  width: 100%;
-}
-
-.bp-textarea {
-  font-family:Roboto, "Helvetica Neue", sans-serif;
-  font-size:14px;
-  letter-spacing:normal;
-  line-height:normal;
-  outline-color:rgb(0, 0, 0);
-  outline-style:none;
-  outline-width:0px;
-  height: auto;
-  margin-bottom: 0.5em;
-}
