+++ src/app/shared/sharer/sharer.component.spec.ts
@@ -1,0 +1,11 @@
+/* tslint:disable:no-unused-variable */
+
+import { TestBed, async } from '@angular/core/testing';
+import { SharerComponent } from './sharer.component';
+
+describe('Component: Sharer', () => {
+  it('should create an instance', () => {
+    let component = new SharerComponent();
+    expect(component).toBeTruthy();
+  });
+});
